(optimize) improve linking strategy
If a config file being linked already exists, previously the bootstrap
script will throw an error and exit. Now if an config file already exists,
the existing file will be backed up and then the new config will be linked.

+++ scripts/bootstrap.fish
@@ -19,7 +19,7 @@ end
 function download
   argparse 'o/output=' -- $argv
 
-  if command -v curl
+  if command -v curl %> /dev/null
     if [ $_flag_verbose ]
         curl --create-dirs -Lo $_flag_o $argv
     else
@@ -32,26 +32,61 @@ function download
 end
 
 function log
-  if [ !$_flag_q ]
+  if [ -z $_flag_q ]
     printf $argv
   end
 end
 
 # install fisher
 log 'Installing fisher...'
-download -o $fish_config_dir/functions/fisher.fish https://git.io/fisher
+if [ -e $fish_config_dir/functions/fisher.fish ]
+  log 'fisher has been already installed on this machine, running fisher self update...'
+  if [ -n $_flag_V ]
+    fisher self-update
+  else
+    fisher self-update %> /dev/null
+else
+  download -o $fish_config_dir/functions/fisher.fish https://git.io/fisher
+end
 
 # link the fishfile and perform update
 log 'Updating fisher plugins...'
-ln -s $DOTFILES/fish/fishfile $fish_config_dir/fishfile
-fisher
+if [ -e $fish_config_dir/fishfile ]
+  log 'Found existing fishfile in your fish config directory. Installing the plugins with `fish install $plugin`.'
+  log 'In this circumstance, your fishfile will be preserved. The symbolic link will not be created.'
+  log 'That means, further changes to the fishfile will not apply to the one included here.'
+  log 'If you want them to sync, you will need to copy your fishfile here and manually link it back.'
+  for $plugin in cat $DOTFILES/fish/fishfile
+    if [ -n $_flag_V ]
+      fisher install $plugin
+    else
+      fisher install $plugin %> /dev/null
+    end
+  end
+else
+  ln -s $DOTFILES/fish/fishfile $fish_config_dir/fishfile
+  if [ -n $_flag_V ]
+    fisher
+  else
+    fisher %> /dev/null
+  end
+end
 
 # subsitute the DOTFILE_PATH_PLACEHOLDER in fish/config.fish to the real dotfiles path and link it
 log 'Linking fish config...'
 sed -n 's/DOTFILE_PATH_PLACEHOLDER/$DOTFILES/gp' $DOTFILES/fish/config.fish
+if [ -e $fish_config_dir/config.fish ]
+  log 'Found existing fish config, backing it up...'
+  mv $fish_config_dir/config.fish $fish_config_dir/config.fish.backup
+end
 ln -s $DOTFILES/fish/config.fish $fish_config_dir/config.fish
 
 # editorconfig
+log 'Linking editorconfig...'
+if [ -e $HOME/.editorconfig ]
+  log 'Found existing editorconfig, backing it up...'
+  mv $HOME/.editorconfig $HOME/.editorconfig.backup
+end
 ln -s $DOTFILES/editorconfig/editorconfig $HOME/.editorconfig
 
 # starship
@@ -65,20 +100,20 @@ if command -v starship %> /dev/null
 end
 
 # node (use flag --china to configure china mirrors for npm/yarn)
-if command -v node
+if command -v node %> /dev/null
   log 'Configuring NPM&Yarn...'
-  if [ $_flag_verbose ]
+  if [ -n $_flag_V ]
     npm install --verbose -g npm yarn
-  else if [ $_flag_q ]
+  else if [ -n $_flag_q ]
     npm install --silent -g npm yarn
   else
     npm install -g npm yarn
   end
 
-  if [ $_flag_china ]
-    if [ $_flag_verbose ]
+  if [ -n $_flag_china ]
+    if [ -n $_flag_V ]
       npm install --verbose -g mirror-config-china
-    else if [ $_flag_q ]
+    else if [ -n $_flag_q ]
       npm install --silent -g mirror-config-china
     else
       npm install -g mirror-config-china
